Moving layout to own file
Changed to a module format to allow import/export
