+++ index.html
@@ -18,6 +18,6 @@
     </div>
     <div class="grid"></div>
     <button id="start">Start</button>
-    <script src="index.js"></script>
+    <script type="module" src="index.js"></script>
   </body>
 </html>
